requirements.txt
# Samantha AI Agent — Dependencies (pinned)

# LLM
google-generativeai>=0.8.0
requests>=2.31.0

# Voice
SpeechRecognition>=3.10.0
faster-whisper>=1.0.0
pyttsx3>=2.90

# Audio
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.26.0

# Browser automation
playwright>=1.45.0

# System automation
pyautogui>=0.9.54

# ML (intent classification)
scikit-learn>=1.5.0

# macOS native (optional)
pyobjc>=10.0; platform_system == "Darwin"

# Config
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
